refactoring project structure
