+++ simitrieve/src/main/java/ml/paulobatista/simitrieve/entity/Language.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package ml.paulobatista.simitrieve;
+package ml.paulobatista.simitrieve.entity;
 
 /**
  * @author paulo
