+++ simitrieve/src/main/java/ml/paulobatista/simitrieve/entity/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author paulo
+ *
+ */
+package ml.paulobatista.simitrieve.entity;
