+++ simitrieve/src/main/java/ml/paulobatista/simitrieve/preprocessing/Preprocessor.java
@@ -1,7 +1,9 @@
 /**
  * 
  */
-package ml.paulobatista.simitrieve;
+package ml.paulobatista.simitrieve.preprocessing;
+
+import ml.paulobatista.simitrieve.entity.Project;
 
 /**
  * @author paulo
